Turns on and off with a constant cycle time, or turns on all the time
This is a working version of the Blinkenlights with two out of three
modes working. It turns the circuit on and off at a relatively short 30
seconds on/20 seconds off cycle, or can be set to keep the circuit on.
It properly displays the indicator light to reflect the current mode,
and modes can be changed with a button. It currently has a lot of debug
logging.